[update] change value only isMove and implement system
