+++ Assets/Scripts/Events/StartEvent.cs
@@ -1,0 +1,17 @@
+using UnityEngine;
+
+public class TestMessagesShow : MonoBehaviour
+{
+    public MessageSystem messageSystem;
+
+    void Start()
+    {
+        string[] storyMessages = new string[] {
+            "操作方法",
+            "AWSDで移動",
+            "スペースで決定"
+        };
+        messageSystem.ShowMessages(storyMessages);
+        
+    }
+}
