+++ Assets/Scripts/GameRoot.cs
@@ -8,10 +8,10 @@ public class GameRoot
     //public static GameRoot I { get; private set; }
 
     // 状態管理用のbool変数
-    public bool isPlay = false;     // 操作中
-    public bool isMessage = false;  // メッセージ表示中
-    public bool isEvent = false;    // イベント実行中
-    public bool isItem = false;     // アイテム表示中
+    public bool isMove = false;     // 操作中
+    //public bool isMessage = false;  // メッセージ表示中
+    //public bool isEvent = false;    // イベント実行中
+    //public bool isItem = false;     // アイテム表示中
 
     public string ThisRole;
