+++ Assets/Scripts/Player/PFPlayerMove.cs
@@ -17,37 +17,40 @@ public class PFPlayerMove : MonoBehaviour
 
     private void Update()
     {
-        //Debug.Log("UpdateTest");
-        Vector2 dir = Vector2.zero;
-        if (Input.GetKey(KeyCode.A))
+        if (GameRoot.I.isMove == true)
         {
-            Debug.Log("Put A");
-            dir.x = -1;
-            animator.SetInteger("Direction", 3);
-        }
-        else if (Input.GetKey(KeyCode.D))
-        {
-            Debug.Log("Put D");
-            dir.x = 1;
-            animator.SetInteger("Direction", 2);
-        }
+            //Debug.Log("UpdateTest");
+            Vector2 dir = Vector2.zero;
+            if (Input.GetKey(KeyCode.A))
+            {
+                Debug.Log("Put A");
+                dir.x = -1;
+                animator.SetInteger("Direction", 3);
+            }
+            else if (Input.GetKey(KeyCode.D))
+            {
+                Debug.Log("Put D");
+                dir.x = 1;
+                animator.SetInteger("Direction", 2);
+            }
 
-        if (Input.GetKey(KeyCode.W))
-        {
-            Debug.Log("Put W");
-            dir.y = 1;
-            animator.SetInteger("Direction", 1);
-        }
-        else if (Input.GetKey(KeyCode.S))
-        {
-            Debug.Log("Put S");
-            dir.y = -1;
-            animator.SetInteger("Direction", 0);
-        }
+            if (Input.GetKey(KeyCode.W))
+            {
+                Debug.Log("Put W");
+                dir.y = 1;
+                animator.SetInteger("Direction", 1);
+            }
+            else if (Input.GetKey(KeyCode.S))
+            {
+                Debug.Log("Put S");
+                dir.y = -1;
+                animator.SetInteger("Direction", 0);
+            }
 
-        dir.Normalize();
-        animator.SetBool("IsMoving", dir.magnitude > 0);
+            dir.Normalize();
+            animator.SetBool("IsMoving", dir.magnitude > 0);
 
-        GetComponent<Rigidbody2D>().linearVelocity = speed * dir;
+            GetComponent<Rigidbody2D>().linearVelocity = speed * dir;
+        }
     }
 }
