+++ Assets/Scripts/UI/StartEvent.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a49d212b868590b4ebef542041657a5e
