--- Assets/Scripts/UI/TestMessagesShow.cs
@@ -1,17 +1,0 @@
-using UnityEngine;
-
-public class TestMessagesShow : MonoBehaviour
-{
-    public MessageSystem messageSystem;
-
-    void Start()
-    {
-        string[] storyMessages = new string[] {
-            "操作方法",
-            "AWSDで移動",
-            "スペースで決定"
-        };
-        messageSystem.ShowMessages(storyMessages);
-        
-    }
-}
