--- Assets/Scripts/UI/TestMessagesShow.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: bf70b7bc216a1c7438bd8ecdbad1c173
